requirements_telegram.txt
telethon==1.34.0
python-binance==1.0.19
python-dotenv==1.0.0
asyncio
logging 